Halfway through the refactoring
